Combine all 4 function files into 1 file
