+++ tests/test_suscat.py
@@ -1,6 +1,5 @@
-from pythcat import suscat
+from pythcat import pythcat
 import pandas as pd
-# import numpy as np
 import pytest
 
 
@@ -12,20 +11,20 @@ def test_inputs():
     df = pd.DataFrame({'x': [1, 2, 11, 22, 8], 'y': [100, 110, 155, 200, 540]})
     # test df input
     with pytest.raises(Exception):
-        suscat.suscat([[1, 2, 3], [1, 2, 3]], [0])
+        pythcat.suscat([[1, 2, 3], [1, 2, 3]], [0])
     # test num input
     with pytest.raises(Exception):
-        suscat.suscat(df, [1], num="")
+        pythcat.suscat(df, [1], num="")
     # test n input
     with pytest.raises(Exception):
-        suscat.suscat(df, [1], n=1.1, num='number')
+        pythcat.suscat(df, [1], n=1.1, num='number')
     with pytest.raises(Exception):
-        suscat.suscat(df, [1], n=20000000, num='number')
+        pythcat.suscat(df, [1], n=20000000, num='number')
     with pytest.raises(Exception):
-        suscat.suscat(df, [1], n=200, num='percent')
+        pythcat.suscat(df, [1], n=200, num='percent')
     # test column input
     with pytest.raises(Exception):
-        suscat.suscat(df, '1')
+        pythcat.suscat(df, '1')
 
 
 def test_outputs():
@@ -35,8 +34,8 @@ def test_outputs():
 
     df = pd.DataFrame({'x': [1, 2, 11, 22, 7], 'y': [100, 110, 120, 130, 140]})
 
-    out1 = suscat.suscat(df, [0, 1], n=80)
-    out2 = suscat.suscat(df, [0], num='number', n=2)
+    out1 = pythcat.suscat(df, [0, 1], n=80)
+    out2 = pythcat.suscat(df, [0], num='number', n=2)
 
     assert isinstance(out1, dict)
     assert len(out1.keys()) == 2 and len(out1[1]) == 4
